backend/requirement.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
python-dotenv==1.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.7
boto3>=1.33,<2.0
PyJWT==2.8.0
flask-cors==4.0.0
gunicorn==21.2.0
